some renames, catch bad uri
