+++ wandb/sdk/launch/runner/loader.py
@@ -27,11 +27,11 @@ def load_backend(
 ) -> AbstractRunner:
     # Static backends
     if backend_name in ["local", "local-container"]:
-        from .local import LocalContainerRunner
+        from .local_container import LocalContainerRunner
 
         return LocalContainerRunner(api, backend_config)
     elif backend_name in ["bare", "local-process"]:
-        from .bare import LocalProcessRunner
+        from .local_process import LocalProcessRunner
 
         return LocalProcessRunner(api, backend_config)
     elif backend_name == "gcp-vertex":
