+++ wandb/sdk/launch/runner/local_process.py
@@ -6,7 +6,7 @@ import wandb
 from wandb.errors import LaunchError
 
 from .abstract import AbstractRun, AbstractRunner
-from .local import _run_entry_point
+from .local_container import _run_entry_point
 from .._project_spec import get_entry_point_command, LaunchProject
 from ..builder.build import get_env_vars_dict
 from ..utils import (
@@ -51,14 +51,18 @@ class LocalProcessRunner(AbstractRunner):
             raise LaunchError("Launch LocalProcessRunner received empty project dir")
 
         # Check to make sure local python dependencies match run's requirement.txt
-        source_entity, source_project, run_name = parse_wandb_uri(launch_project.uri)
-        validate_wandb_python_deps(
-            source_entity,
-            source_project,
-            run_name,
-            self._api,
-            launch_project.project_dir,
-        )
+        try:
+            source_entity, source_project, run_name = parse_wandb_uri(launch_project.uri)
+        except ValueError:
+            _logger.warning(f"Could not parse URI {launch_project.uri} into run name.")
+        else:
+            validate_wandb_python_deps(
+                source_entity,
+                source_project,
+                run_name,
+                self._api,
+                launch_project.project_dir,
+            )
 
         env_vars = get_env_vars_dict(launch_project, self._api)
         for env_key, env_value in env_vars.items():
